Refactor vector.py into different files
